FIX: Parse the trailing digit run with long.TryParse to match the widened counter [UUM-145766]
Parse the trailing numeric suffix as a long so names whose digits exceed Int32.MaxValue seed the counter instead of falling through to the append path.

With namesTried already a long, long.TryParse lets a name like "item2147483648" increment to "item2147483649" on collision, rather than failing int.TryParse and producing "item21474836481". A name is only treated as an unseeded base when its digit run exceeds Int64.MaxValue.

Jira: https://jira.unity3d.com/browse/UUM-145766

+++ Packages/com.unity.inputsystem/InputSystem/Runtime/Utilities/StringHelpers.cs
@@ -337,9 +337,9 @@ namespace UnityEngine.InputSystem.Utilities
                 while (lastDigit > 0 && char.IsDigit(baseName[lastDigit - 1]))
                     --lastDigit;
                 if (lastDigit != baseName.Length &&
-                    int.TryParse(baseName.Substring(lastDigit), out var trailingNumber))
+                    long.TryParse(baseName.Substring(lastDigit), out var trailingNumber))
                 {
-                    namesTried = (long)trailingNumber + 1;
+                    namesTried = trailingNumber + 1;
                     baseName = baseName.Substring(0, lastDigit);
                 }
             }
